+++ pytket/binders/passes.cpp
@@ -526,18 +526,9 @@ PYBIND11_MODULE(passes, m) {
       "SynthesiseTket", &SynthesiseTket,
       "Optimises and converts all gates to CX, TK1 and Phase gates.");
   m.def(
-      "SynthesiseOQC",
-      []() {
-        tket_log()->warn(
-            "SynthesiseOQC is deprecated. It will be removed "
-            "after pytket v1.28.");
-        return SynthesiseOQC();
-      },
-      "Optimises and converts all gates to ECR, Rz, SX and Phase. "
-      "DEPRECATED: will be removed after pytket 1.28.");
-  m.def(
       "SynthesiseUMD", &SynthesiseUMD,
-      "Optimises and converts all gates to XXPhase, PhasedX, Rz and Phase.");
+      "Optimises and converts all gates to XXPhase, PhasedX, Rz and Phase. "
+      "DEPRECATED: will be removed after pytket 1.32.");
   m.def(
       "SquashTK1", &SquashTK1,
       "Squash sequences of single-qubit gates to TK1 gates.");
@@ -846,8 +837,8 @@ PYBIND11_MODULE(passes, m) {
 
   m.def(
       "CliffordResynthesis", &gen_clifford_resynthesis_pass,
-      "An optimisation pass that resynhesises all Clifford subcircuits and "
-      "then applies some rewrite rules to simplify them further."
+      "An optimisation pass that resynthesises Clifford subcircuits, trying "
+      "to reduce the 2-qubit gate count as much as possible."
       "\n\n:param transform: optional user-provided resynthesis method to "
       "apply to all Clifford subcircuits (a function taking a Clifford "
       "circuit as an argument and returning an equivalent circuit); if not "