+++ tket/test/src/test_CompilerPass.cpp
@@ -271,7 +271,7 @@ SCENARIO("Test making (mostly routing) passes using PassGenerators") {
          std::make_shared<LexiRouteRoutingMethod>()});
 
     PassPtr all_passes =
-        SynthesiseOQC() >> SynthesiseUMD() >> SynthesiseTK() >> cp_route;
+        SynthesiseTket() >> SynthesiseUMD() >> SynthesiseTK() >> cp_route;
     REQUIRE(all_passes->apply(cu));
     REQUIRE(cu.check_all_predicates());
   }
@@ -448,7 +448,7 @@ SCENARIO("Construct sequence pass") {
 
 SCENARIO("Construct sequence pass that is invalid in strict mode") {
   std::vector<PassPtr> invalid_pass_to_combo{
-      SynthesiseOQC(), SynthesiseUMD(), SynthesiseTK()};
+      SynthesiseTket(), SynthesiseUMD(), SynthesiseTK()};
   for (const PassPtr& pass : invalid_pass_to_combo) {
     std::vector<PassPtr> passes = {pass};
     OpTypeSet ots = {OpType::CX};